fix(config): set testWriter default to fast, update runtime types + distribute tests to per-op files
